MangaUpdates tweaks + fixing the scorelist from main
Since 0.1-0.9 aren't valid of mangaupdates

Fixes #1324

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/mangaupdates/MangaUpdatesApi.kt
@@ -123,7 +123,7 @@ class MangaUpdatesApi(
     private suspend fun updateSeriesRating(track: Track) {
         if (track.score != 0f) {
             val body = buildJsonObject {
-                put("rating", track.score.toInt())
+                put("rating", track.score)
             }
             authClient.newCall(
                 PUT(
